Introduce restate-platform foundation crate
Part of the effort to decompose the restate-types monolith into focused,
composable utility crates.

Add `restate-platform` as a low-dependency foundation crate that provides
key type aliases, traits, and essential types the rest of the project can
reliably depend upon. It sits at the bottom of the dependency graph with
minimal transitive dependencies.

Moves the following into restate-platform:
- Error traits (GenericError, CodedError) from restate-types
- Storage traits (StorageEncode, StorageDecode) from restate-types
- Network marker trait (NetSerde) from restate-encoding
- Memory estimation trait (EstimatedMemorySize) from restate-memory
- Hash re-exports (HashMap, HashSet) for consistent hashing
- Sync utilities (OwnedSemaphorePermit)
