+++ crates/clock/Cargo.toml
@@ -14,7 +14,8 @@ test-util = []
 
 [dependencies]
 restate-workspace-hack = { workspace = true }
-restate-encoding = { workspace = true }
+
+restate-platform = { workspace = true }
 
 bilrost = { workspace = true }
 jiff = { workspace = true, optional = true }
