+++ crates/encoding/Cargo.toml
@@ -11,13 +11,10 @@ publish = false
 restate-workspace-hack = { workspace = true }
 
 restate-encoding-derive = { version = "0.1.0", path = "derive" }
-restate-sharding = { workspace = true }
+restate-platform = { workspace = true }
 
 bilrost = { workspace = true }
 bytes = { workspace = true }
-bytestring = { workspace = true }
 
 [dev-dependencies]
 rand = { workspace = true }
-restate-sharding = { workspace = true, features = ["bilrost"] }
-static_assertions = { workspace = true }
