+++ crates/encoding/src/lib.rs
@@ -12,97 +12,10 @@ mod bilrost_as;
 pub mod bilrost_encodings;
 mod common;
 
-use std::collections::BTreeMap;
-use std::collections::HashMap;
-use std::collections::HashSet;
-use std::ops::RangeInclusive;
-use std::sync::Arc;
-
 pub use bilrost_as::BilrostAsAdaptor;
 pub use bilrost_encodings::{Arced, ArcedSlice, RestateEncoding};
 pub use common::U128;
 pub use restate_encoding_derive::{BilrostAs, BilrostNewType, NetSerde};
-/// A marker trait for types that can be serialized and sent over the network.
-///
-/// Types implementing this trait are considered eligible for wire transmission,
-/// typically via serialization. It is intended to be implemented automatically
-/// using the `#[derive(NetSerde)]` macro.
-///
-/// # Example
-/// ```ignore
-/// #[derive(NetSerde)]
-/// struct MyMessage {
-///     a: u64,
-///     b: String,
-/// }
-/// ```
-pub trait NetSerde {}
-
-macro_rules! impl_net_serde {
-    ($t:ty) => {
-        impl NetSerde for $t {}
-    };
-    ($($t:ty),+) => {
-        $(impl_net_serde!($t);)+
-    }
-}
-
-impl_net_serde!(
-    bool,
-    usize,
-    u8,
-    u16,
-    u32,
-    u64,
-    u128,
-    isize,
-    i8,
-    i16,
-    i32,
-    i64,
-    i128,
-    String,
-    bytes::Bytes,
-    bytestring::ByteString,
-    std::time::Duration
-);
-
-macro_rules! impl_net_serde_tuple {
-    ($($t:ident),+) => {
-        impl<$($t),+> NetSerde for ($($t),+) where $($t: NetSerde),+ {}
-    };
-}
-
-impl_net_serde_tuple!(T0, T1);
-impl_net_serde_tuple!(T0, T1, T2);
-impl_net_serde_tuple!(T0, T1, T2, T3);
-impl_net_serde_tuple!(T0, T1, T2, T3, T4);
-impl_net_serde_tuple!(T0, T1, T2, T3, T4, T5);
-impl_net_serde_tuple!(T0, T1, T2, T3, T4, T5, T6);
-
-impl<T> NetSerde for Vec<T> where T: NetSerde {}
-impl<T> NetSerde for Option<T> where T: NetSerde {}
-impl<K, V, S> NetSerde for HashMap<K, V, S>
-where
-    K: NetSerde,
-    V: NetSerde,
-{
-}
-
-impl<K, V> NetSerde for BTreeMap<K, V>
-where
-    K: NetSerde,
-    V: NetSerde,
-{
-}
-
-impl<V> NetSerde for HashSet<V> where V: NetSerde {}
-impl<Idx> NetSerde for RangeInclusive<Idx> where Idx: NetSerde {}
-impl NetSerde for restate_sharding::PartitionId {}
-impl<T> NetSerde for Arc<T> where T: NetSerde {}
-impl<T> NetSerde for Arc<[T]> where T: NetSerde {}
-impl<T> NetSerde for Box<T> where T: NetSerde {}
-impl<T, const N: usize> NetSerde for [T; N] where T: NetSerde {}
 
 #[cfg(test)]
 mod test {
@@ -131,50 +44,5 @@ mod test {
         let y = Flattened::decode(bytes).expect("decodes");
 
         assert_eq!(x.id.0, y.id);
-    }
-
-    /// Validates that `KeyRange`'s general bilrost encoding produces the same
-    /// wire format as `RangeInclusive<u64>` with `RestateEncoding`. This ensures
-    /// that `KeyRange` fields can use `#[bilrost(N)]` and remain wire-compatible
-    /// with the old `#[bilrost(tag(N), encoding(RestateEncoding))] RangeInclusive<u64>`.
-    #[test]
-    fn key_range_wire_compat_with_range_inclusive() {
-        use restate_sharding::KeyRange;
-
-        use super::RestateEncoding;
-
-        #[derive(Debug, PartialEq, bilrost::Message)]
-        struct WithKeyRange {
-            #[bilrost(1)]
-            range: KeyRange,
-        }
-
-        #[derive(Debug, PartialEq, bilrost::Message)]
-        struct WithRangeInclusive {
-            #[bilrost(tag(1), encoding(RestateEncoding))]
-            range: std::ops::RangeInclusive<u64>,
-        }
-
-        for (start, end) in [(0u64, 0u64), (1, 100), (0, u64::MAX), (42, 42)] {
-            let kr = KeyRange::new(start, end);
-            let ri = start..=end;
-
-            let kr_bytes = WithKeyRange { range: kr }.encode_to_vec();
-            let ri_bytes = WithRangeInclusive { range: ri.clone() }.encode_to_vec();
-
-            assert_eq!(
-                kr_bytes, ri_bytes,
-                "wire format mismatch for range ({start}, {end})"
-            );
-
-            // Cross-decode
-            let decoded: WithRangeInclusive =
-                WithRangeInclusive::decode(&*kr_bytes).expect("cross-decode KeyRange→RI");
-            assert_eq!(decoded.range, ri);
-
-            let decoded: WithKeyRange =
-                WithKeyRange::decode(&*ri_bytes).expect("cross-decode RI→KeyRange");
-            assert_eq!(decoded.range, kr);
-        }
     }
 }
