+++ crates/memory/Cargo.toml
@@ -13,13 +13,12 @@ default = []
 [dependencies]
 restate-workspace-hack = { workspace = true }
 
+restate-platform = { workspace = true }
 restate-serde-util = { workspace = true }
 
 bytes = { workspace = true }
 futures = { workspace = true }
-hashbrown = { workspace = true }
 metrics = { workspace = true }
-parking_lot = { workspace = true }
 pin-project-lite = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["sync"] }
