+++ crates/memory/src/controller.rs
@@ -8,11 +8,11 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use hashbrown::HashMap;
 use metrics::{Gauge, gauge};
-use parking_lot::RwLock;
 use tracing::warn;
 
+use restate_platform::hash::HashMap;
+use restate_platform::sync::RwLock;
 use restate_serde_util::NonZeroByteCount;
 
 use crate::MemoryPool;
