+++ crates/platform/src/lib.rs
@@ -1,0 +1,26 @@
+// Copyright (c) 2023 - 2026 Restate Software, Inc., Restate GmbH.
+// All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+pub mod errors;
+pub mod hash;
+pub mod memory;
+pub mod network;
+pub mod storage;
+pub mod sync;
+
+// prelude exports;
+pub mod prelude {
+    pub use crate::errors::*;
+    pub use crate::hash::{HashMap, HashSet};
+    pub use crate::memory::EstimatedMemorySize;
+    pub use crate::network::NetSerde;
+    pub use crate::sync::{Mutex, RwLock};
+    pub use restate_util_string::{ReString, ToReString, format_restring};
+}
