+++ crates/platform/src/memory.rs
