+++ crates/platform/src/sync.rs
@@ -1,0 +1,12 @@
+// Copyright (c) 2023 - 2026 Restate Software, Inc., Restate GmbH.
+// All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+pub type RwLock<T> = parking_lot::RwLock<T>;
+pub type Mutex<T> = parking_lot::Mutex<T>;
