+++ crates/sharding/Cargo.toml
@@ -15,6 +15,8 @@ bilrost = ["dep:bilrost"]
 [dependencies]
 restate-workspace-hack = { workspace = true }
 
+restate-platform = { workspace = true }
+
 bilrost = { workspace = true, optional = true }
 derive_more = { workspace = true, features = ["deref", "from", "into", "add", "display", "debug", "from_str"] }
 serde = { workspace = true, optional = true }
