+++ crates/types/src/storage.rs
@@ -16,79 +16,21 @@ use std::sync::Arc;
 
 use bytes::{Buf, BufMut, Bytes, BytesMut};
 use chrono::Utc;
-use downcast_rs::{DowncastSync, impl_downcast};
 
-use restate_encoding::{BilrostAs, NetSerde};
-use restate_memory::EstimatedMemorySize;
+use restate_platform::network::NetSerde;
+// Re-export the traits from restate-platform
+pub use restate_platform::storage::{
+    StorageCodecKind, StorageDecode, StorageDecodeError, StorageEncode, StorageEncodeError,
+};
+
+use restate_encoding::BilrostAs;
+use restate_platform::memory::EstimatedMemorySize;
 use restate_serde_util::ByteCount;
+use restate_util_string::format_restring;
 
-use crate::errors::GenericError;
 use crate::journal_v2::raw::{RawEntry, RawEntryError, TryFromEntry};
 use crate::journal_v2::{Decoder, EntryMetadata, EntryType};
 use crate::time::MillisSinceEpoch;
-
-#[derive(Debug, thiserror::Error)]
-pub enum StorageEncodeError {
-    #[error("encoding failed: {0}")]
-    EncodeValue(GenericError),
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum StorageDecodeError {
-    #[error("failed reading codec: {0}")]
-    ReadingCodec(String),
-    #[error("decoding failed: {0}")]
-    DecodeValue(GenericError),
-    #[error("unsupported codec kind: {0}")]
-    UnsupportedCodecKind(StorageCodecKind),
-}
-
-#[derive(
-    Debug, Copy, Clone, strum::FromRepr, derive_more::Display, PartialEq, Eq, bilrost::Enumeration,
-)]
-#[repr(u8)]
-pub enum StorageCodecKind {
-    /// plain old protobuf
-    Protobuf = 1,
-    /// flexbuffers + serde (length-prefixed)
-    FlexbuffersSerde = 2,
-    /// length-prefixed raw-bytes. length is u32
-    LengthPrefixedRawBytes = 3,
-    // Note: Discriminant 4 was previously `BincodeSerde`, which was added to prepare for future
-    // use but was never actually used for persistent storage. It has been removed because bincode
-    // is no longer maintained. The discriminant is intentionally skipped to prevent accidental
-    // reuse.
-    /// Json (no length prefix)
-    Json = 5,
-    /// Bilrost (no length-prefixed)
-    Bilrost = 6,
-    /// A custom encoding that does not rely on any of the standard encoding formats
-    /// supported by the [`encode`] and [`decode`] modules.
-    ///
-    /// When using this variant, the encoding and decoding logic is entirely defined
-    /// by the implementation of the [`StorageEncode`] and [`StorageDecode`] traits.
-    ///
-    /// While you may still use utility functions from the [`encode`] and [`decode`] modules,
-    /// it is up to your implementation to decide how (or if) to use them, and how the final
-    /// byte representation is constructed.
-    Custom = 7,
-}
-
-impl From<StorageCodecKind> for u8 {
-    fn from(value: StorageCodecKind) -> Self {
-        value as u8
-    }
-}
-
-impl TryFrom<u8> for StorageCodecKind {
-    type Error = StorageDecodeError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        StorageCodecKind::from_repr(value).ok_or_else(|| {
-            StorageDecodeError::ReadingCodec(format!("unknown discriminant '{value}'"))
-        })
-    }
-}
 
 /// Codec which encodes [`StorageEncode`] implementations by first writing the
 /// [`StorageEncode::default_codec`] byte and then encoding the value part via
@@ -119,7 +61,7 @@ impl StorageCodec {
 
     pub fn decode<T: StorageDecode, B: Buf>(buf: &mut B) -> Result<T, StorageDecodeError> {
         if buf.remaining() < mem::size_of::<u8>() {
-            return Err(StorageDecodeError::ReadingCodec(format!(
+            return Err(StorageDecodeError::ReadingCodec(format_restring!(
                 "remaining bytes in buf '{}' < version bytes '{}'",
                 buf.remaining(),
                 mem::size_of::<u8>()
@@ -132,34 +74,6 @@ impl StorageCodec {
         // decode value
         T::decode(buf, codec)
     }
-}
-
-/// Trait to encode a value using the specified [`Self::default_codec`]. The trait is used by the
-/// [`StorageCodec`] to first write the codec byte and then the serialized value via
-/// [`Self::encode`].
-///
-/// # Important
-/// The [`Self::encode`] implementation should use the codec specified by [`Self::default_codec`].
-pub trait StorageEncode: DowncastSync {
-    fn encode(&self, buf: &mut BytesMut) -> Result<(), StorageEncodeError>;
-
-    /// Codec which is used when encode new values.
-    fn default_codec(&self) -> StorageCodecKind;
-}
-impl_downcast!(sync StorageEncode);
-
-static_assertions::assert_obj_safe!(StorageEncode);
-
-/// Trait to decode a value given the [`StorageCodecKind`]. This trait is used by the
-/// [`StorageCodec`] to decode a value after reading the used storage codec.
-///
-/// # Important
-/// To support codec evolution, this trait implementation needs to be able to decode values encoded
-/// with any previously used codec.
-pub trait StorageDecode {
-    fn decode<B: Buf>(buf: &mut B, kind: StorageCodecKind) -> Result<Self, StorageDecodeError>
-    where
-        Self: Sized;
 }
 
 /// Implements the [`StorageEncode`] and [`StorageDecode`] by encoding/decoding the implementing
@@ -279,101 +193,6 @@ impl StorageEncode for PolyBytes {
 }
 
 static_assertions::assert_impl_all!(PolyBytes: Send, Sync);
-
-/// Enable simple serialization of String types as length-prefixed byte slice
-impl StorageEncode for String {
-    fn default_codec(&self) -> StorageCodecKind {
-        StorageCodecKind::LengthPrefixedRawBytes
-    }
-
-    fn encode(&self, buf: &mut BytesMut) -> Result<(), StorageEncodeError> {
-        let my_bytes = self.as_bytes();
-        buf.put_u32_le(u32::try_from(my_bytes.len()).map_err(|_| {
-            StorageEncodeError::EncodeValue(
-                anyhow::anyhow!("only support serializing types of size <= 4GB").into(),
-            )
-        })?);
-        if buf.remaining_mut() < my_bytes.len() {
-            return Err(StorageEncodeError::EncodeValue(
-                anyhow::anyhow!(format!(
-                    "not enough buffer space to serialize value;\
-                        required {} bytes but free capacity was {}",
-                    my_bytes.len(),
-                    buf.remaining_mut()
-                ))
-                .into(),
-            ));
-        }
-        buf.put_slice(my_bytes);
-        Ok(())
-    }
-}
-impl StorageDecode for String {
-    fn decode<B: ::bytes::Buf>(
-        buf: &mut B,
-        kind: StorageCodecKind,
-    ) -> Result<Self, StorageDecodeError>
-    where
-        Self: Sized,
-    {
-        match kind {
-            StorageCodecKind::LengthPrefixedRawBytes => {
-                if buf.remaining() < mem::size_of::<u32>() {
-                    return Err(StorageDecodeError::DecodeValue(
-                        anyhow::anyhow!(
-                            "insufficient data: expecting {} bytes for length",
-                            mem::size_of::<u32>()
-                        )
-                        .into(),
-                    ));
-                }
-                let length = usize::try_from(buf.get_u32_le()).expect("u32 to fit into usize");
-
-                if buf.remaining() < length {
-                    return Err(StorageDecodeError::DecodeValue(
-                        anyhow::anyhow!(
-                            "insufficient data: expecting {} bytes for flexbuffers",
-                            length
-                        )
-                        .into(),
-                    ));
-                }
-
-                let bytes = buf.take(length);
-                Ok(String::from_utf8_lossy(bytes.chunk()).to_string())
-            }
-            codec => Err(StorageDecodeError::UnsupportedCodecKind(codec)),
-        }
-    }
-}
-
-// Enable simple serialization of Bytes types as length-prefixed byte slice
-impl StorageEncode for bytes::Bytes {
-    fn default_codec(&self) -> StorageCodecKind {
-        StorageCodecKind::LengthPrefixedRawBytes
-    }
-
-    fn encode(&self, buf: &mut BytesMut) -> Result<(), StorageEncodeError> {
-        buf.put_u32_le(u32::try_from(self.len()).map_err(|_| {
-            StorageEncodeError::EncodeValue(
-                anyhow::anyhow!("only support serializing types of size <= 4GB").into(),
-            )
-        })?);
-        if buf.remaining_mut() < self.len() {
-            return Err(StorageEncodeError::EncodeValue(
-                anyhow::anyhow!(format!(
-                    "not enough buffer space to serialize value;\
-                        required {} bytes but free capacity was {}",
-                    self.len(),
-                    buf.remaining_mut()
-                ))
-                .into(),
-            ));
-        }
-        buf.put_slice(&self[..]);
-        Ok(())
-    }
-}
 
 /// A marker stored in the storage
 ///
